Predavanje osam

+++ videoteka/app/Http/Controllers/GenreController.php
@@ -6,6 +6,7 @@ use App\Models\Genre;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\App;
+use Exception;
 
 class GenreController extends Controller
 {
@@ -14,21 +15,19 @@ class GenreController extends Controller
      */
     public function index()
     {
-
         $locale = App::currentLocale(); // en/sr
+        
+        //en -> sort po name_en
+        //sr -> sort po name_sr
 
-        //en-> sort po name_en
-        //sr-> sort po name_sr
-
-        if($locale=='en') {
-            $data = Genre::orderBy('name_en')->paginate(5);
-        } elseif($locale=='sr'){
-            $data = Genre::orderBy('name_sr')->paginate(5);
-        } else{
-            // all dovlaci sve podatke iz tabele genres
-            $data = Genre::paginate(5);
-        }           
-
+        if($locale=='en'){
+            $data = Genre::orderBy('name_en')->paginate(4);
+        }elseif($locale=='sr'){
+            $data = Genre::orderBy('name_sr')->paginate(4);
+        }else{
+            //all dovlaci sve podatke iz tabele genres
+            $data = Genre::paginate(4);
+        }        
         return view('genre.index', ['data'=>$data]);
     }
 
@@ -49,23 +48,25 @@ class GenreController extends Controller
             'name_en' => 'required|unique:genres,name_en|alpha',
             'name_sr' => 'nullable|unique:genres,name_sr|alpha'
         ]);
-      
-        // Ovo se izvrsava ukoliko je forma prosla validaciju
+    
 
-        /* 1.create
-        Genre::create(['id'=>5, 'name_en'=>'mistery', 'name_sr'=>'misterija']);
-        2.instanca klase
-        $g = new Genre;
-        $g->name_en = 'mistery';
-        $g->name_sr = 'misterija';
+        //kod ispod se izvrsava u slucaju da je forma prosla validaciju
+
+        //1. create
+        //Genre::create(['name_en'=>'Mystery', 'name_sr'=>'Misterija']); 
+        //2. istanca klase
+        /*$g = new Genre;
+        $g->name_en = 'Mystery';
+        $g->name_sr = 'Misterija';
         $g->save();*/
 
-        Genre::create($request->all());
-        
+        Genre::create($request->all()); 
+
         $request->session()->flash('alertType', 'success');
         $request->session()->flash('alertMsg', 'Successfully added.');
 
         return redirect()->route('genre.index');
+            
     }
 
     /**
@@ -79,8 +80,10 @@ class GenreController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
+    //$genre = Genre::find($request->input('id'))
     public function edit(Genre $genre)
     {
+        //return view('genre.edit', ['genre'=>$genre]);
         return view('genre.edit', compact('genre'));
     }
 
@@ -92,12 +95,12 @@ class GenreController extends Controller
         $request->validate([
             'name_en' => [
                 'required', 
-                //'unique:genres, name_en',
+                //'unique:genres,name_en', 
                 Rule::unique('genres', 'name_en')->ignore($genre->id),
                 'alpha'],
-            'name_sr' => [ 
-                'required', 
-                //'unique:genres, name_sr',
+            'name_sr' => [
+                'nullable', 
+                //'unique:genres,name_sr', 
                 Rule::unique('genres', 'name_sr')->ignore($genre->id),
                 'alpha']
         ]);
@@ -115,11 +118,37 @@ class GenreController extends Controller
      */
     public function destroy(Genre $genre)
     {
-        $genre->delete();
+        try{
+            $genre->delete();
 
-        session()->flash('alertType', 'success');
-        session()->flash('alertMsg', 'Successfully deleted.');
+            session()->flash('alertType', 'success');
+            session()->flash('alertMsg', 'Successfully deleted.');
 
-        return redirect()->route('genre.index');
+            return redirect()->route('genre.index');
+        }
+        catch(Exception $e) {
+            //echo 'Message: ' .$e->getMessage();
+            session()->flash('alertType', 'danger');
+            session()->flash('alertMsg', 'Cannot be deleted.');
+
+            return redirect()->route('genre.index');
+        }
+    }
+
+    public function destroy2(Genre $genre)
+    {
+        if($genre->films->count()){
+            session()->flash('alertType', 'danger');
+            session()->flash('alertMsg', 'Cannot be deleted. Films='.$genre->films->count());
+
+            return redirect()->route('genre.index');
+        }else{
+            $genre->delete();
+    
+            session()->flash('alertType', 'success');
+            session()->flash('alertMsg', 'Successfully deleted.');
+    
+            return redirect()->route('genre.index');        
+        }
     }
 }
